osx hates cleaning up files
